FEAT: Basic Login & Register Modal
