+++ frontend/src/app/layout.tsx
@@ -2,8 +2,10 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { useState } from 'react'
 import { Navigation } from '@/components/Navigation'
-import '../styles/global.css'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+
+// @ts-expect-error: Import global.css
+import '../styles/global.css'
 
 export default function IndexLayout({
   children,
@@ -17,9 +19,17 @@ export default function IndexLayout({
         <body>
           <QueryClientProvider client={queryClient}>
             <div className='flex flex-row h-full w-full'>
+              {/* Navigation Component*/}
               <Navigation/>
-              {children}
+              
+              {/* Page Content */}
+              <main className="w-full h-full">{children}</main>
+
+              {/* Portal Div for Modals */}
               <div id="modal-root"></div>
+
+
+              {/* React Query Dev Tools only for Development */}
               <ReactQueryDevtools initialIsOpen={false} />
             </div>
           </QueryClientProvider>
