--- frontend/src/app/login/layout.tsx
@@ -1,7 +1,0 @@
-export default function LoginLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <section>{children}</section>;
-}
