--- frontend/src/components/ClassButton.tsx
@@ -1,30 +1,0 @@
-import { Home } from "lucide-react";
-
-interface ClassButtonProps {
-  icon?: React.ReactNode;
-  label: string;
-  onclick?: () => void;
-  className?: string;
-}
-
-
-
-export function ClassButton({
-  icon,
-  label,
-  onclick,
-  className,
-}: ClassButtonProps) {
-
-
-  return (
-    <button
-      onClick={onclick}
-      className={"flex items-center rounded-[8] p-2 bg-foreground" + 
-                (className ? " " + className : "")}
-    >
-      <span className="">{label}</span>
-      {icon && <span className="mr-4">{icon}</span>}
-    </button>
-  );
-}
